move core logic files to root dir
